test(sf-364): invert the witness -- the refused op is now refused cleanly
Renames tests/integration-rust/rejected-op-witness.test.ts to
rejected-op-refusal.test.ts and flips all five assertions. The witness
encoded the broken behaviour on purpose so the fix had something concrete
to flip; the old file does not survive the rename, so CI never holds a
test asserting the defect after the fix.

- R1: the refused op leaves both the in-memory queue and the durable
  pending set, and the local value is still readable (keep-and-present).
- R2: confirmWrite answers 'rejected', not 'timeout'.
- R3: a legitimate write queued behind the refused one is acked.
- R4: topgun_client_op_refusals_total{transport="ws",reason="forbidden"}
  rises by exactly one and then stays put across three further writes --
  the refused op is never re-sent. The pre-write read asserts the series
  is present, since an absent line is a broken instrument, not a zero.
- R5: the application receives exactly one rejection for the write,
  carrying a machine-readable cause and the server's reason.
